requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.3
django-autoslug==1.9.9
django-ckeditor==6.6.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-upload-validator==1.1.6
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
